+++ src/webview/main.tsx
@@ -1,6 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { App } from "./App";
+import { ThemeProvider } from "./ThemeProvider";
 import { postWebviewLog } from "./vscode";
 
 const root = document.getElementById("root");
@@ -14,7 +15,9 @@ postWebviewLog("info", "webview bundle loaded");
 try {
   createRoot(root).render(
     <StrictMode>
-      <App />
+      <ThemeProvider>
+        <App />
+      </ThemeProvider>
     </StrictMode>
   );
   postWebviewLog("info", "react app render scheduled");